encrypted-media/drmtoday-temporary-cenc.html

<!doctype html>
<meta name="timeout" content="long">
<html>
  <head>
    <meta charset=utf-8>
    <title>Encrypted Media Extensions ClearKey Success</title>
    <link rel="help" href="https://w3c.github.io/encrypted-media/">

    <script src=/resources/testharness.js></script>
    <script src=/resources/testharnessreport.js></script>
    <script src=/encrypted-media/util/utils.js></script>
    <script src=/encrypted-media/util/fetch.js></script>
    <script src=/encrypted-media/util/testmediasource.js></script>
    <script src=/encrypted-media/util/base64.js></script>
    <script src=/encrypted-media/util/drmtoday-messagehandler.js></script>

    <meta name="timeout" content="long">
  </head>
  <body>
    <meta name="timeout" content="long">
    <h1 class="instructions">Description</h1>
    <p class="instructions">
      This test verifies success paths for the Clear Key keysystem.
    </p>

    <div id='log'></div>

    <div id='video'></div>

    <script>
        var config = {  element:            document.getElementById('video'),
                        keysystem:          getKeySystem(),
                        sessiontype:        'temporary',
                        messagehandler:     messagehandler,
                        audioType:          'audio/mp4;codecs="mp4a.40.2"',
                        videoType:          'video/mp4;codecs="avc1.4d401e"',
                        audioPath:          '/encrypted-media/content/audio_aac-lc_128k_dashinit.mp4',
                        videoPath:          '/encrypted-media/content/video_512x288_h264-360k_dashinit.mp4',
                        initDataType:       'cenc',
                        duration:           5 };

        var testname = 'DrmToday playback; temporary sessiontype; calling setMediaKeys() after setting src attribute.';

        config.configuration = {    initDataTypes: [ config.initDataType ],
                                    audioCapabilities: [ { contentType: config.audioType } ],
                                    videoCapabilities: [ { contentType: config.videoType } ],
                                    sessionTypes: [ config.sessiontype ] };

        async_test( function( test ) {
            window.console.log( testname + " running" );

            var _video,
                _mediaKeys,
                _mediaKeySession,
                _mediaSource,
                _allKeysUsableEvent = false,
                _waitingForKeyEvent = false,
                _canPlayThroughEvent = false,
                _playingEvent = false,
                _timeupdateEvent = false,
                _releaseSequence = false,
                _events = [ ];

            _video = document.createElement('video');
            _video.setAttribute( 'width', '600px' );
            config.element.appendChild( _video );

            function onMessage(event) {
                assert_equals( event.target, _mediaKeySession );
                assert_true( event instanceof window.MediaKeyMessageEvent );
                assert_equals( event.type, 'message');

                assert_any( assert_equals,
                            event.messageType,
                            _releaseSequence    ? [ 'license-release']
                                                : [ 'license-request', 'individualization-request' ] );

                if ( event.messageType !== 'individualization-request' ) {
                    _events.push( event.messageType );
                }

                config.messagehandler( event.messageType, event.message ).then( function( response ) {
                    if ( event.messageType === 'license-request' ) {
                        _events.push( 'license-response' );
                    } else if ( event.messageType === 'license-release' ) {
                        _events.push( 'release-response' );
                    }

                    waitForEventAndRunStep('keystatuseschange', _mediaKeySession, onKeyStatusesChange, test);
                    _mediaKeySession.update( response ).then( function() {
                        _events.push('update');
                    }).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                });
            }

            function onKeyStatusesChange(event) {
                assert_equals(event.target, _mediaKeySession );
                assert_true(event instanceof window.Event );
                assert_equals(event.type, 'keystatuseschange' );

                var hasKeys = false, pendingKeys = false;
                _mediaKeySession.keyStatuses.forEach( function( value, keyid ) {
                    assert_any( assert_equals, value, [ 'status-pending', 'usable' ] );

                    hasKeys = true;
                    pendingKeys = pendingKeys || ( value === 'status-pending' );

                });

                if ( !_allKeysUsableEvent && hasKeys && !pendingKeys ) {
                    _allKeysUsableEvent = true;
                    _events.push( 'allkeysusable' );

                    window.console.log("Set media keys");
                    _video.setMediaKeys(_mediaKeys);
                }

                if ( !hasKeys ) {
                    _events.push( 'emptykeyslist' );
                }
            }

            function onEncrypted(event) {
                assert_equals(event.target, _video);
                assert_true(event instanceof window.MediaEncryptedEvent);
                assert_equals(event.type, 'encrypted');

                waitForEventAndRunStep('message', _mediaKeySession, onMessage, test);
                _mediaKeySession.generateRequest(   event.initDataType, event.initData ).then( function() {
                    _events.push( 'generaterequest' );
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }

            function onClosed(event) {
                window.console.log( testname + ' closed' );
                assert_array_equals( _events,
                                    [
                                        'generaterequest',
                                        'license-request',
                                        'license-response',
                                        'update',
                                        'allkeysusable',
                                        'playing',
                                        'close'
                                    ],
                                    "Expected events sequence" );

                _video.src = "";
                _video.setMediaKeys( null ).then(function(){
                    test.done();
                });
            }

            function onTimeupdate(event) {
                if ( _video.currentTime > ( config.duration || 5 ) && !_timeupdateEvent ) {
                    _timeupdateEvent = true;
                    _video.pause();

                    _mediaKeySession.closed.then( test.step_func( onClosed ) );
                    _mediaKeySession.close().then( function() {
                        _events.push( 'close' );
                    }).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }
            }

            function onPlaying(event) {
                _playingEvent = true;
                _events.push( 'playing' );

                // Not using waitForEventAndRunStep() to avoid too many
                // EVENT(onTimeUpdate) logs.
                _video.addEventListener('timeupdate', onTimeupdate, true);
            }

            navigator.requestMediaKeySystemAccess(config.keysystem, [config.configuration]).then(function(access) {
                return access.createMediaKeys();
            }).then(function(mediaKeys) {
                _mediaKeys = mediaKeys;
            }).then(function() {
                // Create the MediaKeySession
                _mediaKeySession = _mediaKeys.createSession( config.sessiontype );

                waitForEventAndRunStep('encrypted', _video, onEncrypted, test);
                waitForEventAndRunStep('playing', _video, onPlaying, test);
            }).then(function() {
                return testmediasource(config);
            }).then(function(source) {
                _mediaSource = source;
                window.console.log("Set video source attribute");
                _video.src = URL.createObjectURL(_mediaSource);
                _video.play();
            }).catch(function(error) {
                forceTestFailureFromPromise(test, error);
            });
        }, testname);

    </script>
  </body>
</html>
